% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepwalk.R
\name{.slw}
\alias{.slw}
\title{Variable to store the information about the current \code{sleepwalk} session.}
\usage{
.slw
}
\description{
Variable to store the information about the current \code{sleepwalk} session.
}
