\name{sleepwalk-package}
\alias{sleepwalk-package}
\alias{sleepwalk}
\docType{package}
\title{
\packageTitle{sleepwalk}
}
\description{
\packageDescription{sleepwalk}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sleepwalk}
\packageIndices{sleepwalk}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sleepwalk}

Maintainer: \packageMaintainer{sleepwalk}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
