\name{sleepwalkMulti}
\alias{sleepwalkMulti}
\title{ Multi-view variant of sleepwalk}
\description{
Displays several embeddings side-by-side in the style of sleepwalk
}
\usage{
sleepwalkMulti(embeddings, featureMatrices, maxdists)
}
\arguments{
  \item{embeddings}{
a list of several embeddings (each a Nx2 matrix as in 'sleepwalk')
}
  \item{featureMatrices}{
a list of several featureMatrices, corresponding to the embeddings (each a NxM matrix as in 'sleepwalk')
}
  \item{maxdists}{
a vector of several maximum distances (as in sleepwalk)
}
}
\details{
}
\value{
None
}
\references{
}
\author{
Simon Anders
}
\note{
}

\seealso{
sleepwalk
}
\examples{
}