% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepwalk.R
\name{slw_on_selection}
\alias{slw_on_selection}
\title{On selection}
\usage{
slw_on_selection(points, emb)
}
\arguments{
\item{points}{a vector of indices of the selected points.}

\item{emb}{an index of the embedding, where the points have been selected.}
}
\description{
This function is called each time any points are selected or deselected.
You can customise it by redefining.
}
