% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepwalk.R
\docType{data}
\name{.slw}
\alias{.slw}
\title{Variable to store the information about the current \code{sleepwalk} session.}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.slw
}
\description{
Variable to store the information about the current \code{sleepwalk} session.
}
\keyword{datasets}
